Enable users to proceed to 'next' after logging in
If a user is not logged in and attempts to access a page that requires
logging in, the user may log in and proceed directly to the page. This
functionality uses the 'next' query parameter in the URL to proceed,
which 'routes.py' accesses through the use of Flask's 'requests' module.