Create upload form to update profile picture
Form field is in 'account.html' and will list any errors. Encoding type
for this form is set to "multipart/form-data" for file upload. Field is
defined in 'forms.py' as an optional field that only accepts 'jpg' and
'png' files. Submitting a new image will not yet update the profile
picture, but the form field itself is operating.