Enable renaming, saving, and displaying new image
When a user uploads a new profile image, the image is stored under a new
name (an 8 byte hex with the same file extension) within the flaskblog
file system. The image filename is committed to the database so that the
template will refer to it when displaying the profile picture.