Create page for updating posts
New 'update_post' view function allows a user to make updates to an
existing post's title and content. The user must be the post's author,
or access to the post's 'update' page will be denied. A 'GET' request
to the update page for a post will retrieve a form similar to the
'new_post' form, but the form fields will be pre-filled with existing
post title and content. Submitting the form updates the database and
redirects the user to the post's page.