Add page number buttons to 'home.html'
We iterate over posts.iter_pages() to display page numbers as links to
the home page. Each link includes query parameters (used in 'routes.py'
to determine which page and corresponding posts to display) that match
the number displayed in the link. For all 'None' types within
posts.iter_pages(), an ellipses is shown.