Add new forms for requesting/setting new password
RequestResetForm has an email field and a submit field. RequestResetForm
checks that the email exists in the database. Users will submit the form
to request a password-reset email. ResetPasswordForm has two password
fields (one for the new password, one to confirm the new password) and a
submit field. Data in the second password field must match the first.
Users will submit this form to set a new password.