Add 'reset_token' view
New view accepts a token, which was generated by User.get_reset_token()
in 'models.py'. Authenticated users are redirected to 'home'. The token
is verified to confirm which user requested the password reset. URLs
with invalid tokens receive a flash message and are redirected to
'reset_request'. URLs with valid tokens return the 'reset_token'
template.